refactor(instance): split auto_provision.rs god object into modules
